compute reference in double for validation

+++ python_benchmarks/test_layernorm.py
@@ -196,9 +196,9 @@ def test_layernorm_bwd_benchmark(
 
     if not disable_validation:
         eager_output = torch.nn.functional.layer_norm(
-            inputs, inputs.shape[1:], weight=weights, bias=bias
+            inputs.to(torch.double), inputs.shape[1:], weight=weights.to(torch.double), bias=bias.to(torch.double)
         )
-        eager_output.backward(grads)
+        eager_output.backward(grads.to(torch.double))
         fd.validate(
             [inputs, grads, mean, invstd, weights],
             [inputs.grad, weights.grad, bias.grad],
